requirements.txt
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.8.0
keras>=2.8.0
opencv-python>=4.5.0
pillow>=8.3.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
requests>=2.26.0
beautifulsoup4>=4.10.0
selenium>=4.0.0
webdriver-manager>=3.8.0
tqdm>=4.62.0
albumentations>=1.1.0
imgaug>=0.4.0
google-images-search>=1.4.0
bing-image-downloader>=1.1.0
